refactor(model): separate `route`, `route_builder`, `state`, `vehicle`
